# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Reservas y Espacios",
        "description": "Documentación de la API para gestionar reservas y espacios.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener todas las reservas del usuario",
                "operationId": "c18fe40736a295c69322918d3c915ad9",
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Sala de Conferencias A"
                                            },
                                            "reservation_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-03-15"
                                            },
                                            "start_hour": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "14:00"
                                            },
                                            "end_hour": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "17:00"
                                            },
                                            "estimated_cost": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 150.75
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron reservas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva",
                "operationId": "da1a59b124ed2072df98076a3f368a0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "space_id",
                                    "reservation_date",
                                    "start_hour",
                                    "end_hour"
                                ],
                                "properties": {
                                    "space_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-15"
                                    },
                                    "start_hour": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:00"
                                    },
                                    "end_hour": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva creada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "field": [
                                                    "El campo field es obligatorio."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Eliminar una reserva",
                "operationId": "0d2019c6cc0efa3fc6982cc6ee5029cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No puedes cancelar esta reserva ya que falta una hora o menos para empezar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Reserva no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}